# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  github_com_etkecc_mrs_internal_model.MatrixError:
    properties:
      errcode:
        description: Matrix error code, e.g M_UNAUTHORIZED
        type: string
      error:
        description: Matrix error message
        type: string
    type: object
...
  github_com_etkecc_mrs_internal_model.QueryDirectoryResponse:
    properties:
      room_id:
        type: string
      servers:
        items:
          type: string
        type: array
    type: object
  github_com_etkecc_mrs_internal_model.QueryServerKeysRequest:
    properties:
      server_keys:
        additionalProperties: {}
        type: object
    type: object
  github_com_etkecc_mrs_internal_model.RoomDirectoryFilter:
    properties:
      generic_search_term:
        type: string
      room_types:
        items:
          type: string
        type: array
    type: object
  github_com_etkecc_mrs_internal_model.RoomDirectoryRequest:
    properties:
      filter:
        $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomDirectoryFilter'
      limit:
        type: integer
      since:
        type: string
    type: object
  github_com_etkecc_mrs_internal_model.RoomDirectoryResponse:
    properties:
      chunk:
        items:
          $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomDirectoryRoom'
        type: array
      next_batch:
        type: string
      prev_batch:
        type: string
      total_room_count_estimate:
        type: integer
    type: object
  github_com_etkecc_mrs_internal_model.RoomDirectoryRoom:
    properties:
      avatar_url:
        type: string
      canonical_alias:
        type: string
      guest_can_join:
        type: boolean
      join_rule:
        type: string
      name:
        type: string
      num_joined_members:
        type: integer
      room_id:
        type: string
      room_type:
        type: string
      topic:
        type: string
      world_readable:
        type: boolean
    type: object
...
  github_com_etkecc_mrs_internal_model.ServerKeysQueryResponse:
    properties:
      server_keys:
        description: each element is a signed ServerKeys object
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
    type: object
...
info:
  contact: {}
  description: Matrix Rooms Search crawls the Matrix federation and serves public-room
    search. It also answers part of the Matrix federation and client-server APIs for
    the rooms it has indexed. Where our implementation diverges from the spec, that
    divergence is documented here on purpose, not hidden. matrixrooms.info is a public
    demo running MRS.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0
  title: Matrix Rooms Search API
  version: 0.0.0
paths:
...
  /_matrix/federation/v1/publicRooms:
    get:
      consumes:
      - application/json
      description: 'Our slice of the federation public-rooms directory: the rooms
        we have crawled and indexed. Authenticated federation endpoint, so a missing
        or invalid X-Matrix signature is a 401. GET (query params) and POST (JSON
        filter body) share one handler, and a malformed POST body is logged then ignored,
        so you get the unfiltered listing rather than an error.'
      parameters:
      - description: Filter and pagination, POST only
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomDirectoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Our indexed slice of the public-rooms directory
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomDirectoryResponse'
        "401":
          description: Federation auth failed (missing or invalid X-Matrix signature)
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      security:
      - FederationAuth: []
      summary: Public rooms directory
      tags:
      - matrix-s2s
    post:
      consumes:
      - application/json
      description: 'Our slice of the federation public-rooms directory: the rooms
        we have crawled and indexed. Authenticated federation endpoint, so a missing
        or invalid X-Matrix signature is a 401. GET (query params) and POST (JSON
        filter body) share one handler, and a malformed POST body is logged then ignored,
        so you get the unfiltered listing rather than an error.'
      parameters:
      - description: Filter and pagination, POST only
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomDirectoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Our indexed slice of the public-rooms directory
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.RoomDirectoryResponse'
        "401":
          description: Federation auth failed (missing or invalid X-Matrix signature)
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      security:
      - FederationAuth: []
      summary: Public rooms directory
      tags:
      - matrix-s2s
  /_matrix/federation/v1/query/directory:
    get:
      description: 'Resolves a room alias to a room ID and the servers that can help
        you join. This is an authenticated federation endpoint: we check the incoming
        X-Matrix signature first, so a missing or bad one is a 401 before any resolution
        happens. Past that we pass back whatever status the lookup produced, so a
        404 means we genuinely could not resolve it, not that the endpoint is missing.'
      parameters:
      - description: 'Room alias to resolve, e.g. #room:server'
        in: query
        name: room_alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room ID and the servers that can help you join
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.QueryDirectoryResponse'
        "401":
          description: Federation auth failed (missing or invalid X-Matrix signature)
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
        "404":
          description: Alias could not be resolved
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      security:
      - FederationAuth: []
      summary: Resolve a room alias
      tags:
      - matrix-s2s
...
  /_matrix/key/v2/query:
    post:
      consumes:
      - application/json
      description: 'Notary lookup for a batch of servers, each returned verbatim as
        its origin signed it. Malformed JSON gets a 400. An empty batch (valid body,
        no server_keys) gets a 200 with an empty key set: you asked for nothing, you
        got nothing.'
      parameters:
      - description: Servers and key IDs to look up
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_etkecc_mrs_internal_model.QueryServerKeysRequest'
      - description: Only return keys valid until at least this timestamp (ms). A
          malformed value is treated as 0.
        in: query
        name: minimum_valid_until_ts
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Signed server keys, or an empty key set for an empty batch
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.ServerKeysQueryResponse'
        "400":
          description: Request body is not valid JSON
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Batch query server keys
      tags:
      - matrix-s2s
  /_matrix/key/v2/query/{serverName}:
    get:
      description: Notary lookup of one remote server's signing keys, handed back
        verbatim as the server signed them (we do not re-wrap or re-sign). A server
        we cannot reach or verify yields an empty key set, not a 5xx tantrum.
      parameters:
      - description: Server name to look up keys for
        in: path
        name: serverName
        required: true
        type: string
      - description: Only return keys valid until at least this timestamp (ms). A
          malformed value is treated as 0.
        in: query
        name: minimum_valid_until_ts
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Signed server keys, or an empty key set
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.ServerKeysQueryResponse'
      summary: Query one server's keys
      tags:
      - matrix-s2s
...
produces:
- application/json
schemes:
- https
securityDefinitions:
  AdminAuth:
    type: basic
  CatalogAuth:
    type: basic
  DiscoveryAuth:
    type: basic
  FederationAuth:
    description: 'Matrix federation signature auth: an `Authorization: X-Matrix origin=...,key=...,sig=...`
      header carrying an ed25519 signature over the request, which we verify against
      the origin server''s published keys.'
    in: header
    name: Authorization
    type: apiKey
  MetricsAuth:
    type: basic
  ModerationAuth:
    type: basic
swagger: "2.0"